Separate scalar metrics table calculation into a separate utils file.
